fix bug crash after clear memory

+++ app/src/main/java/net/iGap/module/FileUtils.java
@@ -515,7 +515,7 @@ public class FileUtils {
      * @param fileOrDirectory main iGap directory
      */
     public static void deleteRecursive(File fileOrDirectory) {
-        if (fileOrDirectory.isDirectory()) {
+        if (fileOrDirectory.isDirectory() && fileOrDirectory.listFiles() != null) {
             for (File child : fileOrDirectory.listFiles()) {
                 deleteRecursive(child);
             }
@@ -693,8 +693,9 @@ public class FileUtils {
     }
 
     private void clearFile(@NotNull File fileTmp) {
-        for (File file : Objects.requireNonNull(fileTmp.listFiles())) {
-            if (!file.isDirectory()) file.delete();
-        }
+        if (fileTmp.listFiles() != null)
+            for (File file : fileTmp.listFiles()) {
+                if (!file.isDirectory()) file.delete();
+            }
     }
 }
